+++ data-raw/baystudy.R
@@ -35,14 +35,14 @@ boattow_baystudy<-read_csv(file.path("data-raw", "Baystudy", "BoatTow.csv"),
                            col_types = cols_only(Year="i", Survey="i", Station="c", Time="c", Tide="i"))%>%
   mutate(Time=parse_date_time(Time, orders="%m/%d/%Y %H:%M:%S"))%>%
   group_by(Year, Survey, Station)%>%
-  filter(Time==min(Time))%>% # Just keep records at the time of the first tow
+  filter(Time==min(Time))%>% # Just keep tide and time-of-day records at the time of the first tow
   ungroup()%>%
   left_join(tidecodes_baystudy, by="Tide")%>%
   select(-Tide)%>%
   rename(Tidetow=Description)
 
 baystudy <- left_join(boattow_baystudy, boatstation_baystudy, by=c("Year", "Survey", "Station"))%>%
-  mutate(Tide=if_else(is.na(Tidetow), Tidestation, Tidetow),
+  mutate(Tide=if_else(is.na(Tidetow), Tidestation, Tidetow), # If tide data were not collected at the time of the tow, use the value from the overall station visit
          Datetime=parse_date_time(if_else(is.na(Time), NA_character_, paste0(Date, " ", hour(Time), ":", minute(Time))), "%Y-%m-%d %H:%M", tz="America/Los_Angeles"),
          Source="Baystudy")%>%
   left_join(baystudy_stations, by="Station")%>%